chore: complete phases 0–2, CI lint fixes, dual-hero images
Phase 0/1: program route audit passes (0 issues).
Phase 2: platform-control-plane index, ai-studio-charter rename, Container tab
precedence banner, Dev Studio imports updated.
Phase 5: fix 28 PLATFORM_DEFAULTS single-quote leaks (targeted ESLint line fix).
Nav: apprentice, case manager, mentor, program catalog in Portals menu.
HeroPicture: responsive desktop/mobile images; media-dimensions program hero 16:9.
Add scripts/fix-platform-defaults-eslint-lines.mjs for safe CI repairs.
