+++ apps/admin/app/admin/email-marketing/campaigns/new/NewCampaignClient.tsx
@@ -235,7 +235,7 @@ export default function NewCampaignPage() {
                           HTMLInputElement | HTMLSelectElement | HTMLTextAreaElement
                         >,
                       ) => setCampaign({ ...campaign, subject: e.target.value })}
-                      placeholder="e.g., Welcome to ${PLATFORM_DEFAULTS.orgName}!"
+                      placeholder={`e.g., Welcome to ${PLATFORM_DEFAULTS.orgName}!`}
                       className="w-full px-4 py-2 border border-slate-300 rounded-lg focus:ring-2 focus:ring-brand-blue-500 focus:border-transparent"
                     />
                     <p className="mt-1 text-sm text-black">
