+++ apps/admin/app/admin/program-holders/verification/[id]/review/actions.ts
@@ -107,8 +107,8 @@ export async function submitVerificationDecision(
 
     const subject =
       decision === 'approved'
-        ? 'Your Program Holder Application Has Been Approved — ${PLATFORM_DEFAULTS.orgName}'
-        : 'Update on Your Program Holder Application — ${PLATFORM_DEFAULTS.orgName}';
+        ? `Your Program Holder Application Has Been Approved — ${PLATFORM_DEFAULTS.orgName}`
+        : `Update on Your Program Holder Application — ${PLATFORM_DEFAULTS.orgName}`;
 
     const html =
       decision === 'approved'
