+++ lib/chat/scripts.ts
@@ -162,13 +162,13 @@ Need help with any of these steps?`,
         trigger: 'complaint',
         action: 'transfer_to_human',
         message:
-          'I understand you have a concern. Let me connect you with our support team who can help resolve this. You can also reach them directly at ${PLATFORM_DEFAULTS.supportPhone}.',
+          `I understand you have a concern. Let me connect you with our support team who can help resolve this. You can also reach them directly at ${PLATFORM_DEFAULTS.supportPhone}.`,
       },
       {
         trigger: 'refund',
         action: 'show_contact',
         message:
-          'For refund inquiries, please contact our finance team at info@${PLATFORM_DEFAULTS.canonicalDomain} or call ${PLATFORM_DEFAULTS.supportPhone}.',
+          `For refund inquiries, please contact our finance team at info@${PLATFORM_DEFAULTS.canonicalDomain} or call ${PLATFORM_DEFAULTS.supportPhone}.`,
       },
       {
         trigger: 'emergency',
@@ -287,7 +287,7 @@ Try answering them, and I can check your work!`,
         trigger: 'technical_issue',
         action: 'redirect_to_page',
         message:
-          'For technical issues, please visit our support page or contact info@${PLATFORM_DEFAULTS.canonicalDomain}',
+          `For technical issues, please visit our support page or contact info@${PLATFORM_DEFAULTS.canonicalDomain}`,
         target: '/support',
       },
     ],
