+++ lib/email/templates/platform-emails.ts
@@ -52,7 +52,7 @@ ${PLATFORM_DEFAULTS.orgName}
    * EMAIL 2: Licensing Proposal
    */
   licensingProposal: {
-    from: 'partnerships@${PLATFORM_DEFAULTS.canonicalDomain}',
+    from: `partnerships@${PLATFORM_DEFAULTS.canonicalDomain}`,
     subject: 'Platform licensing proposal',
     getHtml: (data: { name: string }) => `
       <div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto; padding: 20px;">
