+++ lib/navigation.ts
@@ -243,11 +243,15 @@ export const NAV_ITEMS: NavItem[] = [
       { name: '— Sign in —', href: '/portals', isHeader: true },
       { name: 'Portal hub', href: '/portals' },
       { name: 'Student / learner', href: '/login?redirect=/learner/dashboard' },
+      { name: 'Apprentice', href: '/login/apprentice' },
       { name: 'Employer', href: '/login?redirect=/employer/dashboard' },
       { name: 'Instructor', href: '/login?redirect=/instructor/dashboard' },
       { name: 'Partner', href: '/login?redirect=/partner/dashboard' },
       { name: 'Program holder', href: '/login?redirect=/program-holder/dashboard' },
+      { name: 'Case manager', href: '/login?redirect=/case-manager/dashboard' },
+      { name: 'Mentor', href: '/login?redirect=/mentor/dashboard' },
       { name: 'Staff', href: '/login?redirect=/admin/staff-portal/dashboard' },
+      { name: 'Program catalog', href: '/programs/catalog' },
 
       { name: '— Career support —', href: '/career-services', isHeader: true },
       { name: 'Career services', href: '/career-services' },
